src/client/requirements.txt
grpcio-tools==1.45.0
Flask==2.1.2
